+++ lib/common/consts/settings.dart
@@ -17,8 +17,8 @@ EndpointData networkEndpointKusama = EndpointData.fromJson(const {
 EndpointData networkEndpointAcala = EndpointData.fromJson(const {
   'info': 'acala-mandala',
   'ss58': 42,
-  'text': 'Acala Mandala (Hosted by Polkawallet)',
-  'value': 'wss://39.99.168.67/wss',
+  'text': 'Acala Mandala (Hosted by Acala Network)',
+  'value': 'wss://testnet-node-1.acala.laminar.one/ws',
 });
 
 List<EndpointData> networkEndpoints = [
@@ -52,20 +52,14 @@ List<EndpointData> networkEndpoints = [
   EndpointData.fromJson(const {
     'info': 'acala-mandala',
     'ss58': 42,
-    'text': 'Mandala TC3 Node 1 (Hosted by OnFinality)',
+    'text': 'Mandala TC4 Node 1 (Hosted by OnFinality)',
     'value': 'wss://node-6684611762228215808.jm.onfinality.io/ws'
   }),
   EndpointData.fromJson(const {
     'info': 'acala-mandala',
     'ss58': 42,
-    'text': 'Mandala TC3 Node 2 (Hosted by OnFinality)',
+    'text': 'Mandala TC4 Node 2 (Hosted by OnFinality)',
     'value': 'wss://node-6684611760525328384.rz.onfinality.io/ws'
-  }),
-  EndpointData.fromJson(const {
-    'info': 'acala-mandala',
-    'ss58': 42,
-    'text': 'Acala Mandala (Hosted by Acala Network)',
-    'value': 'wss://testnet-node-1.acala.laminar.one/ws',
   }),
 ];
 
@@ -85,4 +79,4 @@ const String acala_stable_coin = 'AUSD';
 const String acala_stable_coin_view = 'aUSD';
 
 /// test app versions
-const String app_beta_version = '0.8.2-beta.1';
+const String app_beta_version = '0.8.3-beta.1';