+++ lib/service/substrateApi/apiAccount.dart
@@ -85,6 +85,7 @@ class ApiAccount {
       webApi.assets.fetchBalance();
     }
     if (store.settings.endpoint.info == networkEndpointAcala.info) {
+      store.acala.setTransferTxs([], reset: true);
       store.acala.loadCache();
     } else {
       // refresh user's staking info if network is kusama or polkadot